requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
databricks-connect==14.1.0
databricks-sdk==0.12.0
Faker==19.12.0
faker-pyspark==0.8.0
faker_airtravel==0.4
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
numpy==1.26.1
pandas==2.1.2
protobuf==4.24.4
py4j==0.10.9.7
pyarrow==13.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
